+++ Dockerfile
@@ -1,0 +1,8 @@
+FROM node:16
+
+WORKDIR /code
+
+COPY package.json /code/package.json
+COPY package-lock.json /code/package-lock.json
+
+RUN npm install
